Add title handling methods for length
Titles may exceed the MediaWiki length limit.
We want to gracefully handle that, thus we introduce
a title formatting in three parts : a fixed front part,
a variable, and a fixed rear. Both fixed parts are kept as it.
The variable part is splitted based on its spaces, and as much
as possible is kept to fit the available length.