Revamp post-pocessing method wrapping signature
We assumed in the past the post processing methods
had to be related to a MetadataMapper, which may not be the case
and is actually quite cumbersome − for example, a regex method
to search and replace patterns need not to be.

Here we implement a workaround for that, where we can either pass
the name of the MetadataMapper method OR an actual method (with
arguments by the way).
Ultimately, the reliance on the MetadataMapper should go.