Revamp metadata mapping process
In previous commit, we revamped the way the metadata enrichment
was done to allow for more flexibility. The direct metadata mapping
had to follow this, and become decoupled from the metadata collection
(which made no sense whatsoever anyway).

Move lot of code to PostProcessing module.

This is a heavily breaking change.