Add Processing method to split and apply a template.
As a developer, I want to transform a list of values to
be wrapped in a Commons template - such as Technique.

Add a general method which splits a string following a given
separator, translate each bit using a given dictionary, and wraps
each in a given template.
Add method implementing the above for {{Technique}}.