Move verifyDescription to bot constructor
The bot may be fired in 'freewheel' mode, where it does not ask to check
the description at each upload. This is of course essential when doing
any sort of actual uploading.

The problem is that this is decided in the overloaded _doUpload method,
which is not meant to be accessible at the library level
(it is called by run() but never directly).

Add a new parameter to the bot constructor, which defaults as True.