Use flat structure for Record XML export
As an uploader, I want to have a flat XML,
in order to use it with the GWToolset.

In 69e95b0, where we introduced XML serialisation,
we chose to export sets as XML subelements.

But the GWToolset can only deal with flat XMLs.
If there is a nested structure, all subelements
are concatenated, deafeating the all purpose.

Here we change the handling of sets so that, if given
a field 'X' with a set of N elements, it is serialized XML
as N elements <X_1></X_1>...<X_N></X_N>.

The obvious drawback is that the field 'X' may not contain the same
number of elements for every record. Unsure how the GWToolset would
deal with that.