Safeguard against DIMS parsing errors using Exception
One of the built-in post-processors is process_DIMS,
which parses a string Joconde-style to extract the values and units.

This parsing may fail if we encounter a slightly different format.
This manifest by a ValueError exception, as the typecasting to float
would fail on a string.

Here we surround this all operation by a try statement, catching a
ValueError. If so, the old field value is returned (so nothing is changed)