Reduce maximum length when cutting title
MediaWiki has a maximum title file length, which means
that the title may have to be cut according to some rules.

I thought this limit was 256 bytes but it turns out to be
actually 240 bytes (as hinted by the GWToolset).

- Change max length to 240
- Add assertion to ensure it

Note how the limits are in bytes and not characters,
but Python len() method on Unicode strings actually
behaves as expected.